move inline sanitizer logic specific to amp mustache
