+++ extensions/amp-mustache/0.1/amp-mustache.js
@@ -15,15 +15,48 @@
  */
 
 import {dict} from '../../../src/utils/object';
+import {htmlSanitizer} from '../third_party/caja/html-sanitizer';
 import {iterateCursor, templateContentClone} from '../../../src/dom';
 import {parse as mustacheParse, render as mustacheRender,
   setUnescapedSanitizier} from '../../../third_party/mustache/mustache';
-import {sanitizeHtml, sanitizeTagsForAmpMustache} from '../../../src/sanitizer';
+import {sanitizeHtml} from '../../../src/sanitizer';
 
 // Configure sanitizer for output of "triple-mustache";a set of allowed tags
 // to be unescaped.
-setUnescapedSanitizier(sanitizeTagsForAmpMustache);
+setUnescapedSanitizier(AmpMustache.sanitizeTagsForTripleMustache_);
 
+/** @const {!Array<string>} */
+const WHITELISTED_TAGS = [
+  'a',
+  'b',
+  'br',
+  'caption',
+  'colgroup',
+  'code',
+  'del',
+  'div',
+  'em',
+  'i',
+  'ins',
+  'mark',
+  'p',
+  'q',
+  's',
+  'small',
+  'span',
+  'strong',
+  'sub',
+  'sup',
+  'table',
+  'tbody',
+  'time',
+  'td',
+  'th',
+  'thead',
+  'tfoot',
+  'tr',
+  'u',
+];
 
 /**
  * Implements an AMP template for Mustache.js.
@@ -33,6 +66,45 @@ setUnescapedSanitizier(sanitizeTagsForAmpMustache);
  * @extends {BaseTemplate$$module$src$service$template_impl}
  */
 export class AmpMustache extends AMP.BaseTemplate {
+
+  /**
+   * Sanitizes user provided HTML to mustache templates, used in amp-mustache.
+   * WARNING: This method should not be used elsewhere as we do not strip out
+   * the style attribute in this method for the inline-style experiment.
+   * We do so in sanitizeHtml which occurs after this initial sanitizing.
+   *
+   * @private
+   * @param {string} html
+   * @return {string}
+   */
+  sanitizeTagsForTripleMustache_(html) {
+    /**
+     * Tag policy for handling what is valid html in templates.
+     * @type {!Function}
+     */
+    const tagPolicy = function(tagName, attribs) {
+      if (tagName == 'template') {
+        for (let i = 0; i < attribs.length; i += 2) {
+          if (attribs[i] == 'type' && attribs[i + 1] == 'amp-mustache') {
+            return {
+              tagName,
+              attribs: ['type', 'amp-mustache'],
+            };
+          }
+        }
+      }
+      const isWhitelistedTag = WHITELISTED_TAGS.includes(tagName);
+      if (!isWhitelistedTag) {
+        return null;
+      }
+      return {
+        tagName,
+        attribs,
+      };
+    };
+
+    return htmlSanitizer.sanitizeWithPolicy(html, tagPolicy);
+  }
 
   /** @override */
   compileCallback() {
